backend/pyproject.toml
[project]
name = "waves-backend"
version = "0.1.0"
description = "Python backend for Waves — local meeting transcription and summarization"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "huggingface-hub>=0.20",
    "transformers>=5.3.0",
    "torch>=2.10.0",
    "accelerate>=1.13.0",
]

[project.optional-dependencies]
local = [
    "faster-whisper>=1.0",
    "llama-cpp-python>=0.2",
]
huggingface = [
    "transformers>=4.40",
    "torch>=2.0",
    "accelerate>=0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["waves"]
